# internal/database/db/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.30.0

package db

import (
	"github.com/jackc/pgx/v5/pgtype"
)

type PrReviewer struct {
	PullRequestID string             `json:"pull_request_id"`
	ReviewerID    string             `json:"reviewer_id"`
	AssignedAt    pgtype.Timestamptz `json:"assigned_at"`
}

type PullRequest struct {
	ID        string             `json:"id"`
	Name      string             `json:"name"`
	AuthorID  string             `json:"author_id"`
	Status    string             `json:"status"`
	CreatedAt pgtype.Timestamptz `json:"created_at"`
	MergedAt  pgtype.Timestamptz `json:"merged_at"`
}

type Team struct {
	Name string `json:"name"`
}

type User struct {
	ID       string `json:"id"`
	Username string `json:"username"`
	TeamName string `json:"team_name"`
	IsActive bool   `json:"is_active"`
}